requirements.txt
aiohttp==3.6.2
asn1crypto==1.2.0
async-timeout==3.0.1
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cryptography==2.7
discord.py==1.3.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
gspread==3.1.0
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauth2client==4.1.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2019.3
requests==2.22.0
rsa==4.0
schedule==0.6.0
six==1.12.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.6
websockets==6.0
yarl==1.3.0
